몬스터 공격패턴 완료.
콜리젼 시작딜레이 + 캐릭터와의 충돌처리 + 몬스터가 타겟잡는방식 수정요망.

+++ CommonLib/CommonLib/data_reader/data_reader_manager.h
@@ -106,8 +106,8 @@ public:
 	{
 		std::wstring kfm;
 		float scale;
-		int32_t walk_speed;
-		int32_t run_speed;
+		float walk_speed;
+		float run_speed;
 	};
 	struct StatAtt
 	{
@@ -127,9 +127,9 @@ public:
 	};
 	struct Distance
 	{
-		int32_t sight;
-		int32_t sight_height_up;
-		int32_t sight_height_down;
+		float sight;
+		float sight_height_up;
+		float sight_height_down;
 	};
 	struct Collision
 	{
